backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
pillow==10.3.0  # For image manipulation (e.g., resizing, validation)
pytest==8.2.2
pytest-asyncio==0.23.6
transformers==4.42.3  # Hugging Face Transformers library
torch==2.3.1          # PyTorch framework (for AI model)